feat(benchmarks): add hmmt26 and imo_answerbench benchmarks
- Add HMMT26 adapter (evalscope/hmmt_feb_2026, 33 problems)
- Add IMO-AnswerBench adapter (evalscope/imo-answerbench, 400 problems, LLM judge)
- Refactor hmmt25/hmmt26 into shared evalscope/benchmarks/hmmt/ directory
- Extract common answer extraction helpers into hmmt/utils.py
- Add unit tests for both benchmarks
- Generate docs and _meta JSON via docs-pipeline
